// src/enums/story.enum.ts
export enum EStoryStatusLabel {
  NOT_PUBLIC = "Ẩn truyện",
  NOT_COMPLETED = "Chưa hoàn thành",
  COMPLETED = "Hoàn thành",
}

export enum EStoryStatusKey {
  NOT_PUBLIC = 0,
  NOT_COMPLETED = 1,
  COMPLETED = 2,
}

export enum EChapterStatusLabel {
  NOT_QUALIFY = "Chưa đủ điều kiện",
  QUALIFIED = "Đủ điều kiện",
}

export enum EChapterStatusKey {
  NOT_QUALIFY = 0,
  QUALIFIED = 1,
}